add a command for a fast feature extractor.
The build_fast_feature_extractor command updates the feature extractor,
so that PaMpeR ignore unrelevant assertions when generating recommendations.

This feature is not tested yet.

If the fast feature extractor is not build, PaMpeR uses the slow version instead,
while emitting a warning message.

I also cleaned the code base.