minor changes.
We noticed that checking some properties of types on values of type "node"
is harder than I expected originally.

It might be a good idea to develop a function that takes a path in One_Table
and returns a path to reach the corresponding node in the original syntax tree
of type "Term.term".