SeLFiE: improvement on Deep.
Deep should happen before Sugar because we want to have Dive_In and
Dive_Deeper within other expressions.

This is more than syntactic sugar, and the interpreter has to be
able to access other language construct (Lambdas, Applies, and so on)
while defining what to do with Dive_In and Dive_Deeper.