SeLFiE: 100x faster implementation of the SeLFiE interpreter.
We stopped constructing path-to-unode tables at every single atomic
evaluation.

Now we introduced a generic type for path:
  datatype gen_path  = Inner_Path of UN.inner_path | Outer_Path of outer_path;

and a structure and signature (Gen_Path_To_Unode: GEN_PATH_TO_UNODE) to manage
the gen_path in an integrated fashion.

Due to this change, we no longer use a ML-functor to produce two Eval_Path:
we need only one evaluation function on generic paths.

We might be able to make it even faster by
- building a table of path-to-unode tables in Eval_Unary,
- avoiding building print-to-paths tables too many times in Quaifier_Domain.